+++ src/main/java/org/jsoup/UncheckedIOException.java
@@ -6,7 +6,7 @@ import java.io.IOException;
  * @deprecated Use {@link java.io.UncheckedIOException} instead. This class acted as a compatibility shim for Java
  * versions prior to 1.8.
  */
-// todo annotate @Deprecated in next release (after previous @Deprecations clear)
+@Deprecated
 public class UncheckedIOException extends java.io.UncheckedIOException {
     public UncheckedIOException(IOException cause) {
         super(cause);